requirements.txt
Django
django-cors-headers
django-mysql
django-psdb-engine
djangorestframework
mysql-connector-python
mysqlclient
python-dotenv
sqlparse
cffi
argon2-cffi
argon2-cffi-bindings
dj-database-url
django-environ
pycodestyle
python-decouple
setuptools
whitenoise
gunicorn
psycopg2-binary
dj-static